+++ codebase/apps/mdv_utils/src/PrintMdv/PrintMdv.cc
@@ -747,7 +747,7 @@ int PrintMdv::_getVolume(DsMdvx *mdvx)
   }
 
   if (_params.remap_z_to_constant_grid) {
-    for (int ii = 0; ii < mdvx->getNFields(); ii++) {
+    for (size_t ii = 0; ii < mdvx->getNFields(); ii++) {
       MdvxField *field = mdvx->getField(ii);
       if (field) {
         field->remapVlevels(_params.remap_z_grid.nz,
